Move agent details and edit card out into reusable components on web app
